Last of the reformat patches
- Reformat test utils maven module
- Reformat cli maven module
- Reformat few files in core which wasn't caught the first time
- Reformat classes in hudson util maven module

For whitespace there should now only be a few instances
here and there. There is still a bunch of files which has like one or two
errors, but whitespace and formatting is no longer the predominant error in
sonar.

This is intended as the catch the last patch before M4 hardening begins

Change-Id: I9e9d5fabd1a3340e3819ec6edb84ecb6cff8152c
Signed-off-by: Henrik Lynggaard Hansen <henrik@hlyh.dk>