Performance improvements aimed at reducing number of jobs and runs held in memory.
Jobs are now decorated with a loader that loads the delegate only on
request.

Builds are held in a weakly referenced cache by each Job. A mediator is
used to provide traversal between build objects without actually loading
them until needed.

Builds refer to their previous and next via the mediator. Summary data
is constructed by the mediator and persisted to disk in a _runmap.xml file in
the jobs/<name>/builds directory.

API that return real build objects in Hudson still do so to preserve
compatibility with 'instanceof' checks, but use is depreciated. A
BuildHistory interface has been added that should be used instead.
Jelly scripts for home page and Build History page switched to use
BuildHistory instead of loading builds.

Signed-off-by: Roy Varghese <rovarghe@gmail.com>