Add a placeholder for top level items that cannot be read/loaded.
If for some reason a job config.xml gets deleted or corrupted while Hudson is still running, it can prevent all the jobs from displaying making hudson unusable. This fix returns a placeholder that displays the name and the cause of the error in a tooltip, so that an admin can shut down Hudson safely and fix the error at an appropriate time.

Signed-off-by: Roy Varghese <rovarghe@gmail.com>