Prevent a build load failure from losing all build history for a job.
With this change, a placeholder build object will be created if
a build.xml file fails to load for whatever reason. This ensure user
is able to list some of the cached details of the build (prior to it
becoming unloadable) as well as navigate through it to previous and
next builds.

Signed-off-by: Roy Varghese <rovarghe@gmail.com>